added manifest file support with DJANGO_STATIC_USE_MANIFEST_FILE flag
added manifest file support with DJANGO_STATIC_USE_MANIFEST_FILE flag to save _FILE_MAP content in a file (manifest.json). This is for the use case where we want to upload the css and javascript files manually to a CDN. On production, django-static will directly pick the file names from manifest.json itself instead of doing all the calculations.